+++ jest.config.js
@@ -9,5 +9,14 @@ module.exports = {
   'transform': {
     '^.+\\.js$': '<rootDir>/node_modules/babel-jest',
     '.*\\.(vue)$': '<rootDir>/node_modules/jest-vue-preprocessor'
-  }
+  },
+  'collectCoverage': true,
+  'coverageDirectory': './coverage/',
+  'collectCoverageFrom': [
+    '**/atoms/*/*.vue',
+    '**/molecules/*/*.vue',
+    '**/organisms/*/*.vue',
+    '**/templates/*/*.vue',
+    '**/pages/*/*.vue'
+  ]
 }