+++ src/03-modules/header/Header.stories.js
@@ -9,7 +9,6 @@ import AMiniCart from '../../03-modules/mini-cart/MiniCart.vue'
 import AWishlist from '../../03-modules/wishlist/Wishlist.vue'
 import AQuantityUpdate from '../../molecules/quantity-update/QuantityUpdate.vue'
 import AModal from '../../molecules/modal/Modal.vue'
-import AProductItem from '../../03-modules/product-item/ProductItem.vue'
 import AOffCanvasSidebar from '../../03-modules/off-canvas-sidebar/OffCanvasSidebar.vue'
 
 import menu from './mocks/menu'
@@ -17,7 +16,7 @@ import sideMenu from '../../../mocks/slide-menu'
 import products from '../../../mocks/products'
 import totals from '../../../mocks/totals'
 
-// todo: add modal component to story
+// todo: add product item component to story
 
 storiesOf('Modules/Header', module)
   .addDecorator(StoryRouter())
@@ -30,8 +29,7 @@ storiesOf('Modules/Header', module)
       AMiniCart,
       AWishlist,
       AQuantityUpdate,
-      AModal,
-      AProductItem
+      AModal
     },
     data: () => ({
       menu,
@@ -41,11 +39,12 @@ storiesOf('Modules/Header', module)
       quantity: {
         text: 'Qty:',
         label: ''
-      }
+      },
+      isModalVisible: false
     }),
     methods: {
       showRegister () {
-        this.$refs.modalRegister.show('register')
+        this.isModalVisible = true
       },
       showMiniCart () {
         this.$refs.modalMiniCart.show('mini-cart')
@@ -70,8 +69,8 @@ storiesOf('Modules/Header', module)
           @goToAccount="showRegister"
         />
         <a-modal
-          name="register"
-          ref="modalRegister"
+          :visible="isModalVisible"
+          @close="isModalVisible = false"
           heading="Registred Customers"
         >
           <a-login
@@ -100,44 +99,7 @@ storiesOf('Modules/Header', module)
             return-to-shopping-button="Return to shopping"
           >
             <template #products>
-              <a-product-item
-                v-for="product in products"
-                :key="product.id"
-                :id="product.id"
-                :name="product.name"
-                :url="product.url"
-                :image="product.image"
-                :price="product.price"
-                :special-price="product.specialPrice"
-                :old-price="product.oldPrice"
-                :options="product.options"
-                remove-button="Remove button"
-                item-tag="li"
-                class="minicart__product-item"
-                @remove="removeMethod(product.id)"
-              >
-                <template #quantity>
-                  <a-quantity-update
-                    :value="3"
-                    class="product-item__qty"
-                    :input-id="'qty' + product.id"
-                    decrement-aria-label="Decrease the quantity"
-                    decrement-icon-title="Minus mark"
-                    increment-aria-label="Increase the quantity"
-                    increment-icon-title="Plus mark"
-                    @update="changeMethod"
-                  >
-                    <template #label>
-                      <a-label
-                        class="product-item__qty-label"
-                        :for="'qty' + product.id"
-                      >
-                        Qty:
-                      </a-label>
-                    </template>
-                  </a-quantity-update>
-                </template>
-               </a-product-item>
+              <!-- todo: add ProductListItem here -->
             </template>
            </a-mini-cart>
          </a-off-canvas-sidebar>
@@ -146,7 +108,9 @@ storiesOf('Modules/Header', module)
           ref="modalWishlist"
           heading="Wishlist"
         >
-          <a-wishlist :products="products" />
+          <a-wishlist>
+            <!-- todo: add ProductListItem here -->
+          </a-wishlist>
          </a-off-canvas-sidebar>
       </div>
     `