+++ CHANGELOG.md
@@ -9,20 +9,35 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Vue Info table generator helper
 - badge: tag knob and new test regarding custom tag
 - Tests for radio button
+- Test coverage
+- Codecov badge
+- Tests for Price component
+- Tests for Preformatted component
+- Tests for Textarea component
+
+### Removed
+- -webkit- in transitions
+- Modifiers in collapse component
 
 ### Changed
+- Move ProductItem to atomic design, rename to ProductListItem
+- Move CatalogGridItem to atomic design, rename to ProductGridItem
 - heading: update heading class and add unit test and knobs
 - paragraph: to update stories - add text knobs, update component class name - add prefix
 - container: add props with tag, update tests
 - Update readme
 - Update tests in heading component, remove extends from styles
-- image: class name - add prefix, add basic tests
+- Image: class name - add prefix, add basic tests
 - Refactor styles in checkbox component
 - Refactor styles for radio button
 - Remove unnecessary variables in quantity update styles
+- Components structure
+- Price component (removed props, changed classes, removed useless storeies)
+- Collapse style
 
 ### Fixed
 - Typo in Message.selectors.json file
+- Transitions
 
 ## [0.4.0] - 04-06-2019
 ### Added